Added an attribute 'rootElement' to the DocumentDefinition. This means that the DocumentParser will look at the rootElement before it looks at the documentName. If it can't find a valid rootElement, the documentName is used.
This is convenient because of the following reason:
If you have one web service call that can return different information depending on parameters, you may want to store the different information in different documents. This was not possible without the rootElement attribute because the documentName must match to the rootElement that the web service call returned. Documentnames need to be unike.

The example:
Webservice returns this:
<ServerResponse>
  <ServerResult/>
<ServerResponse>

The name in documentDefinition need to match to the rootElement, Like this:
<Document name="ServerResponse" dataManager="MBRestServiceDataHandler" autoCreate="TRUE">

With the rootElement attribute this is possible:
<Document name="SecondServerResponse" dataManager="MBRestServiceDataHandler" autoCreate="TRUE" rootElement="ServerResponse">

See NNEKNAC-24 for more information why this was needed. This now follows convention that is implemented in the web-project.